+++ backend/Models/Account.cs
@@ -7,7 +7,7 @@ namespace api.Models
   {
     public int Id { get; set; }
     public string Name { get; set; } = String.Empty;
-    public long CurrentBalance { get; set; }
+    public decimal CurrentBalance { get; set; }
     public string UserId { get; set; } = String.Empty;
     public User? User { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.Now;
